ci: add typo checking
typos is a powerful source code spell checker.

Adds a CI job that runs on every PR and
push to master (but can also be run manually with
workflow_dispatch) that checks for typos.

Adds a config file .typos.toml that deals with
false positives and only checks for top-level/one-level
.mediawiki and .md files.